% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{geo_to_h3}
\alias{geo_to_h3}
\title{Coords to H3}
\usage{
geo_to_h3(data, lat, lon, resolution = 7L)
}
\arguments{
\item{data}{A data.frame, \link[tibble]{tibble} containing
\code{lat} and \code{lon}.}

\item{lat, lon}{Coordinates to bin.}

\item{resolution}{Hexbin resolution, see 
\href{https://uber.github.io/h3/#/documentation/core-library/resolution-table}{official documentation}.}
}
\value{
The \code{data} as \link[tibble]{tibble} with their corresponding
hexagon index as \code{hex} column.
}
\description{
Turn coordinates into hexagon indices.
}
\examples{
geo_to_h3(quakes, lat, long)

}
