% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_base_cell}
\alias{get_base_cell}
\title{Base Cell}
\usage{
get_base_cell(data, ...)
}
\arguments{
\item{data}{A data.frame or character vector,
if the former is specified then the \code{hex}
argument \emph{must} be passed, specifying the
bare column name containing hexagon indices.}

\item{...}{Other arguments passed to methods.}
}
\value{
A vector or original data.frame as 
\link[tibble]{tibble} with an additional column 
as specified by \code{.name} which defaults to 
\code{base_cell}.
}
\description{
Get the number of the base cell for a given H3 index.
}
