% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3-assertions.R
\name{h3_is_class_res}
\alias{h3_is_class_res}
\title{Check Class III Resolution}
\usage{
h3_is_class_res(data, ...)
}
\arguments{
\item{data}{A data.frame or character vector,
if the former is specified then the \code{hex}
argument \emph{must} be passed, specifying the
bare column name containing hexagon indices.}

\item{...}{Other arguments passed to methods.}
}
\value{
A vector or original data.frame as 
\link[tibble]{tibble} with an additional column 
as specified by \code{.name} which defaults to 
\code{is_pentagon}.
}
\description{
Whether the given H3 index is in a Class III resolution 
(rotated versus the icosahedron and subject to shape 
distortion adding extra points on icosahedron edges, 
making them not true hexagons).
}
\examples{
h3_is_class_res("invalid")

}
