% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3-assertions.R
\name{h3_is_hex}
\alias{h3_is_hex}
\title{Check Hexagon}
\usage{
h3_is_hex(data, ...)
}
\arguments{
\item{data}{A data.frame or character vector,
if the former is specified then the \code{hex}
argument \emph{must} be passed, specifying the
bare column name containing hexagon indices.}

\item{...}{Other arguments passed to methods.}
}
\value{
A vector or original data.frame as 
\link[tibble]{tibble} with an additional column 
as specified by \code{.name} which defaults to 
\code{is_hex}.
}
\description{
Check whether hexagons are valid.
}
\examples{
h3_is_hex("invalid")
h3_is_hex("809bfffffffffff")

geo_to_h3(quakes, lat, long, resolution = 3L) \%>\% 
  h3_is_hex(hex = hex, .name = "is_it_a_valid_hex")

}
