% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3-assertions.R
\name{h3_is_pentagon}
\alias{h3_is_pentagon}
\title{Check Pentagon}
\usage{
h3_is_pentagon(data, ...)
}
\arguments{
\item{data}{A data.frame or character vector,
if the former is specified then the \code{hex}
argument \emph{must} be passed, specifying the
bare column name containing hexagon indices.}

\item{...}{Other arguments passed to methods.}
}
\value{
A vector or original data.frame as 
\link[tibble]{tibble} with an additional column 
as specified by \code{.name} which defaults to 
\code{is_pentagon}.
}
\description{
Check whether pentagons are valid.
}
\examples{
h3_is_pentagon("invalid")

}
