% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3-assertions.R
\name{h3_is_valid}
\alias{h3_is_valid}
\title{Check Hexagon}
\usage{
h3_is_valid(data, ...)
}
\arguments{
\item{data}{A data.frame or character vector,
if the former is specified then the \code{hex}
argument \emph{must} be passed, specifying the
bare column name containing hexagon indices.}

\item{...}{Other arguments passed to methods.}
}
\value{
A vector or original data.frame as 
\link[tibble]{tibble} with \code{is_valid}
column.
}
\description{
Check whether hexagons are valid.
}
\examples{
h3_is_valid("invalid")

}
