% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{h3_line}
\alias{h3_line}
\title{Get Line Between Hexagons}
\usage{
h3_line(origin, destination)
}
\arguments{
\item{origin, destination}{H3indices to draw line.}
}
\value{
A list of h3indices drawing line.
}
\description{
Given two H3 indexes, return the line of indexes between them (inclusive).
}
\details{
This function may fail to find the line between two indexes,
for example if they are very far apart. It may also fail when finding 
distances for indexes on opposite sides of a pentagon.
}
\note{
Lines are drawn in grid space, and may not correspond exactly to 
either Cartesian lines or great arcs.
}
\examples{
\dontrun{
hexagons <- geo_to_h3(quakes, lat, long) 
h3_line(sample(hexagons$hex, 1), sample(hexagons$hex, 1))
}

}
