% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{h3_to_children}
\alias{h3_to_children}
\title{Get Children}
\usage{
h3_to_children(data, ...)
}
\arguments{
\item{data}{A data.frame or character vector,
if the former is specified then the \code{hex}
argument \emph{must} be passed, specifying the
bare column name containing hexagon indices.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named list of children where the name of the list item
is the corresponding h3index.
}
\description{
Get the children/descendents of the given hexagon at a particular resolution.
}
\examples{
\dontrun{h3_to_children("809bfffffffffff")}

}
