% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{h3_to_geo}
\alias{h3_to_geo}
\title{Get Hexagon Center Coordinates}
\usage{
h3_to_geo(data, hex)
}
\arguments{
\item{data}{A data.frame or tibble containing \code{lat} and \code{lon}.}

\item{hex}{Column containing hexagon indices.}
}
\value{
The \code{data} as \link[tibble]{tibble} with their corresponding
hexagon centers as \code{hex_center_lat} and \code{hex_center_lon} columns.
}
\description{
Get Hexagon Center Coordinates
}
\examples{
geo_to_h3(quakes, lat, long) \%>\% 
  h3_to_geo(hex)

}
