% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{h3_to_geo_boundary}
\alias{h3_to_geo_boundary}
\title{Get Hexagon Boundaries}
\usage{
h3_to_geo_boundary(data, hex)
}
\arguments{
\item{data}{A data.frame or tibble containing \code{lat} and \code{lon}.}

\item{hex}{Column containing hexagon indices.}
}
\value{
A named list of hexagon boundaries where the names are the hex index.
}
\description{
Get Hexagon Boundaries
}
\examples{
geo_to_h3(quakes, lat, long) \%>\% 
  h3_to_geo_boundary(hex)

}
