% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{h3_to_parent}
\alias{h3_to_parent}
\title{Get Parent}
\usage{
h3_to_parent(data, ...)
}
\arguments{
\item{data}{A data.frame or character vector,
if the former is specified then the \code{hex}
argument \emph{must} be passed, specifying the
bare column name containing hexagon indices.}

\item{...}{Other arguments passed to methods.}
}
\value{
A vector or original data.frame as
\link[tibble]{tibble} with an additional column
as specified by \code{.name} which defaults to
\code{parent}.
}
\description{
Get the parent of the given hexagon at a particular resolution.
}
\examples{
h3_to_parent("809bfffffffffff")

}
