% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{h3_to_resolution}
\alias{h3_to_resolution}
\title{Base Resolution}
\usage{
h3_to_resolution(data, ...)
}
\arguments{
\item{data}{A data.frame or character vector,
if the former is specified then the \code{hex}
argument \emph{must} be passed, specifying the
bare column name containing hexagon indices.}

\item{...}{Other arguments passed to methods.}
}
\value{
A vector or original data.frame as
\link[tibble]{tibble} with an additional column
as specified by \code{.name} which defaults to
\code{resolution}.
}
\description{
Get the resolution of an H3 index.
}
\examples{
h3_to_resolution("809bfffffffffff")

}
