% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3inr-package.R
\docType{package}
\name{h3inr-package}
\alias{h3inr}
\alias{h3inr-package}
\title{h3inr: Hexagon-based Geographic Grid System}
\description{
Hexagon-based Geographic Grid System.
}
\author{
\strong{Maintainer}: John Coene \email{jcoenep@gmail.com} (0000-0002-6637-4107)

}
\keyword{internal}
