% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{h3_area}
\alias{h3_area}
\alias{h3_edge_length}
\alias{h3_num_hexagons}
\title{Get Hexagons Information}
\usage{
h3_area(resolution = 4L, unit = c("m2", "km2"))

h3_edge_length(resolution = 4L, unit = c("m", "km"))

h3_num_hexagons(resolution = 4L)
}
\arguments{
\item{resolution}{Hexagon resolution.}

\item{unit}{Metric, either meters (\code{m2})
or kilometers (/code{km}, \code{km2}).}
}
\description{
Average hexagon area, or edge length at a given resolution.

The total count of hexagons in the world at a given resolution. 
Note that above resolution 8 the exact count cannot be represented 
in a JavaScript 32-bit number, so consumers should use caution when 
applying further operations to the output.
}
\examples{
h3_area(7, "km2")
h3_edge_length(7, "km")

h3_num_hexagons(2)

}
