% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{k_ring}
\alias{k_ring}
\title{Get neighbours}
\usage{
k_ring(data, hex, size = 1L)
}
\arguments{
\item{data}{A data.frame or tibble containing \code{lat} and \code{lon}.}

\item{hex}{Column containing hexagon indices.}

\item{size}{Radius of the ring ($k$).}
}
\value{
A named list of neighbours where the names are the hex index
of the which the neighbours have been computed.
}
\description{
Get neighbouring hexagons.
}
\examples{
hexagons <- geo_to_h3(quakes, lat, long)
k_ring(hexagons, hex)

}
