% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{polyfill}
\alias{polyfill}
\title{Polyfill}
\usage{
polyfill(polygon, resolution = 4L)
}
\arguments{
\item{polygon}{A list defining coordiantes of polygon.}

\item{resolution}{Hexbin resolution, see 
\href{https://uber.github.io/h3/#/documentation/core-library/resolution-table}{official documentation}.}
}
\description{
Get the set of hexagons within a polygon.
}
\examples{
# define a polygon
polygon <- list(
 list(37.813318999983238, -122.4089866999972145),
 list(37.7198061999978478, -122.3544736999993603),
 list(37.8151571999998453, -122.4798767000009008)
)

polyfill(polygon, 7L)

}
