% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{k_ring}
\alias{k_ring}
\title{Get neighbours}
\usage{
k_ring(data, size = 1L, ...)
}
\arguments{
\item{data}{A data.frame or character vector,
if the former is specified then the \code{hex}
argument \emph{must} be passed, specifying the
bare column name containing hexagon indices.}

\item{size}{Radius of the ring ($k$).}

\item{...}{Other arguments passed to methods.}
}
\value{
A named list of neighbours where the names are the hex index
of the which the neighbours have been computed.
}
\description{
Get neighbouring hexagons.
}
